+++ dask/dataframe/io/hdf.py
@@ -14,7 +14,9 @@ from ..core import DataFrame, new_dd_object
 from ... import multiprocessing
 from ...base import tokenize, compute_as_if_collection
 from ...bytes.utils import build_name_function
-from ...compatibility import PY3
+from ...compatibility import PY3, _HAVE_PATHLIB, unicode
+if _HAVE_PATHLIB is True:
+    from ...compatibility import pathlib
 from ...context import _globals
 from ...delayed import Delayed, delayed
 from ...local import get_sync
@@ -345,9 +347,9 @@ def read_hdf(pattern, key, start=0, stop=None, columns=None,
 
     Parameters
     ----------
-    pattern : string, list
-        File pattern (string), buffer to read from, or list of file
-        paths. Can contain wildcards.
+    pattern : string, pathlib.Path, list
+        File pattern (string), pathlib.Path, buffer to read from, or list of
+        file paths. Can contain wildcards.
     key : group identifier in the store. Can contain wildcards
     start : optional, integer (defaults to 0), row number to start at
     stop : optional, integer (defaults to None, the last row), row number to
@@ -395,6 +397,8 @@ def read_hdf(pattern, key, start=0, stop=None, columns=None,
         lock = get_scheduler_lock()
 
     key = key if key.startswith('/') else '/' + key
+    if _HAVE_PATHLIB and isinstance(pattern, pathlib.Path):
+        pattern = unicode(pattern)
     if isinstance(pattern, str):
         paths = sorted(glob(pattern))
     else: